# apps/server/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "ApiResponse_null_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "anyOf": [],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "ApiResponse_string_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "Song": {
        "properties": {
          "id": {
            "type": "string"
          },
          "albumId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "codec": {
            "type": "string"
          },
          "hasDolbyAtmos": {
            "type": "boolean"
          },
          "trackNumber": {
            "type": "number",
            "format": "double"
          },
          "discNumber": {
            "type": "number",
            "format": "double"
          },
          "artists": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "composers": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "fileSrc": {
            "type": "string"
          }
        },
        "required": [
          "albumId",
          "title",
          "codec",
          "hasDolbyAtmos",
          "trackNumber",
          "discNumber",
          "artists",
          "composers",
          "duration",
          "fileSrc"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_Song_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Song"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateSongDTO": {
        "properties": {
          "title": {
            "type": "string"
          },
          "codec": {
            "type": "string"
          },
          "hasDolbyAtmos": {
            "type": "boolean"
          },
          "trackNumber": {
            "type": "number",
            "format": "double"
          },
          "discNumber": {
            "type": "number",
            "format": "double"
          },
          "artists": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "composers": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "fileSrc": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse__content-string--language-string_-Array_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "items": {
              "properties": {
                "language": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                }
              },
              "required": ["language", "content"],
              "type": "object"
            },
            "type": "array",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "AddLyricsDTO": {
        "properties": {
          "songId": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": ["songId", "language", "content"],
        "type": "object",
        "additionalProperties": false
      },
      "SongUrlDTO": {
        "properties": {
          "filePath": {
            "type": "string"
          },
          "expiresIn": {
            "anyOf": [
              {
                "type": "number",
                "format": "double"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": ["filePath"],
        "type": "object",
        "additionalProperties": false
      },
...
      "Server": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "httpPort": {
            "type": "number",
            "format": "double"
          },
          "httpsPort": {
            "type": "number",
            "format": "double"
          },
          "tunnelEnabled": {
            "type": "boolean"
          },
          "tunnelUrl": {
            "type": "string"
          },
          "httpsEnabled": {
            "type": "boolean"
          },
          "sslCertPath": {
            "type": "string"
          },
          "sslKeyPath": {
            "type": "string"
          },
          "sslPassword": {
            "type": "string"
          },
          "customUrl": {
            "type": "string"
          },
          "proxyHosts": {
            "type": "string"
          },
          "forceHttps": {
            "type": "boolean"
          },
          "allowRemoteConnections": {
            "type": "boolean"
          },
          "remoteIpFilter": {
            "type": "string"
          },
          "remoteIpFilterMode": {
            "type": "string"
          },
          "enableAutoPortMapping": {
            "type": "boolean"
          },
          "publicHttpPort": {
            "type": "number",
            "format": "double"
          },
          "publicHttpsPort": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "id",
          "name",
          "httpPort",
          "httpsPort",
          "tunnelEnabled",
          "httpsEnabled",
          "forceHttps",
          "allowRemoteConnections",
          "remoteIpFilterMode",
          "enableAutoPortMapping",
          "publicHttpPort",
          "publicHttpsPort"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_Server_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Server"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateServerDTO": {
        "properties": {
          "name": {
            "type": "string"
          },
          "httpPort": {
            "type": "number",
            "format": "double"
          },
          "httpsPort": {
            "type": "number",
            "format": "double"
          },
          "tunnelEnabled": {
            "type": "boolean"
          },
          "tunnelUrl": {
            "type": "string"
          },
          "httpsEnabled": {
            "type": "boolean"
          },
          "sslCertPath": {
            "type": "string"
          },
          "sslKeyPath": {
            "type": "string"
          },
          "sslPassword": {
            "type": "string"
          },
          "customUrl": {
            "type": "string"
          },
          "proxyHosts": {
            "type": "string"
          },
          "forceHttps": {
            "type": "boolean"
          },
          "allowRemoteConnections": {
            "type": "boolean"
          },
          "remoteIpFilter": {
            "type": "string"
          },
          "remoteIpFilterMode": {
            "type": "string"
          },
          "enableAutoPortMapping": {
            "type": "boolean"
          },
          "publicHttpPort": {
            "type": "number",
            "format": "double"
          },
          "publicHttpsPort": {
            "type": "number",
            "format": "double"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "ServerConfigResponse": {
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {}
        },
        "type": "object",
        "additionalProperties": {}
      },
      "ApiResponse_ServerConfigResponse_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ServerConfigResponse"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "ServerConfigDTO": {
        "properties": {
          "autoScan": {
            "type": "boolean"
          },
          "autoScanPeriod": {
            "type": "string"
          },
          "generateChapters": {
            "type": "string"
          },
          "autoSelectTracks": {
            "type": "boolean"
          },
          "preferAudioLan": {
            "type": "string"
          },
          "preferSubsLan": {
            "type": "string"
          },
          "subsMode": {
            "type": "string"
          },
          "tempTranscodeFolder": {
            "type": "string"
          },
          "transcodeBuffer": {
            "type": "number",
            "format": "double"
          },
          "transcodePreset": {
            "type": "string"
          },
          "maxTranscodeProcesses": {
            "type": "number",
            "format": "double"
          },
          "automaticUpdates": {
            "type": "boolean"
          }
        },
        "required": [
          "autoScan",
          "autoScanPeriod",
          "generateChapters",
          "autoSelectTracks",
          "preferAudioLan",
          "preferSubsLan",
          "subsMode",
          "tempTranscodeFolder",
          "transcodeBuffer",
          "transcodePreset",
          "maxTranscodeProcesses",
          "automaticUpdates"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_ServerConfigDTO_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ServerConfigDTO"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateServerConfigDTO": {
        "properties": {
          "autoScan": {
            "type": "boolean"
          },
          "autoScanPeriod": {
            "type": "string"
          },
          "generateChapters": {
            "type": "string"
          },
          "autoSelectTracks": {
            "type": "boolean"
          },
          "preferAudioLan": {
            "type": "string"
          },
          "preferSubsLan": {
            "type": "string"
          },
          "subsMode": {
            "type": "string"
          },
          "tempTranscodeFolder": {
            "type": "string"
          },
          "transcodeBuffer": {
            "type": "number",
            "format": "double"
          },
          "transcodePreset": {
            "type": "string"
          },
          "maxTranscodeProcesses": {
            "type": "number",
            "format": "double"
          },
          "automaticUpdates": {
            "type": "boolean"
          }
        },
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Full user authentication with library and remote access checks"
      },
      "cookieAuthFast": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Fast user authentication without additional checks"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Admin user authentication"
      },
      "managementAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Management access for local or admin users"
      },
      "public": {
        "type": "apiKey",
        "in": "header",
        "name": "x-public-access",
        "description": "Public access without authentication"
      }
    }
  },
  "info": {
    "title": "Seerial Media Server API",
    "version": "0.2.0",
    "description": "Media management server API for the Seerial suite",
    "license": {
      "name": "GPL-3.0-only"
    },
    "contact": {
      "name": "Juan Llavero"
    }
  },
  "paths": {
...
    "/songs/{id}": {
      "patch": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Song_"
                }
              }
            }
          }
        },
        "description": "Update song details",
        "tags": ["Songs"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSongDTO"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "Delete",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Delete a song",
        "tags": ["Songs"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/songs/{id}/lyrics": {
      "get": {
        "operationId": "GetSongsLyrics",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse__content-string--language-string_-Array_"
                }
              }
            }
          }
        },
        "description": "Get song lyrics",
        "tags": ["Songs"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/songs/lyrics": {
      "post": {
        "operationId": "AddSongsLyrics",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_string_"
                }
              }
            }
          }
        },
        "description": "Add song lyrics",
        "tags": ["Songs"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLyricsDTO"
              }
            }
          }
        }
      }
    },
    "/songs/stream-url": {
      "post": {
        "operationId": "GetSongUrl",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_string_"
                }
              }
            }
          }
        },
        "description": "Generate a JWT-signed URL for direct song streaming",
        "tags": ["Songs"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isWeb",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isDesktop",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isMobile",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SongUrlDTO"
              }
            }
          }
        }
      }
    },
    "/songs/stream": {
      "get": {
        "operationId": "StreamAudio",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "description": "Stream audio file",
        "tags": ["Songs"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "isWeb",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/playlists/{id}/songs/{songId}": {
      "delete": {
        "operationId": "RemoveSong",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Remove song from playlist",
        "tags": ["Playlists"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "songId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/servers/{id}": {
      "patch": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Server_"
                }
              }
            }
          }
        },
        "description": "Update server configuration",
        "tags": ["Servers"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerDTO"
              }
            }
          }
        }
      }
    },
    "/servers/config/{key}": {
      "get": {
        "operationId": "GetServerConfigKey",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ServerConfigResponse_"
                }
              }
            }
          }
        },
        "description": "Get a specific server config setting",
        "tags": ["Servers"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/servers/config": {
      "get": {
        "operationId": "GetServerConfig",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ServerConfigDTO_"
                }
              }
            }
          }
        },
        "description": "Get all server config settings",
        "tags": ["Servers"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": []
      },
      "patch": {
        "operationId": "UpdateServerConfig",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ServerConfigDTO_"
                }
              }
            }
          }
        },
        "description": "Update server config settings",
        "tags": ["Servers"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerConfigDTO"
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:34200/api"
    }
  ]
}